⚡ 배송 상태 변경, Chucks 아이템 처리자로 데이터 받아오기

+++ backend/reslow/src/main/java/com/ssafy/reslow/domain/order/dto/OrderComfirmationResponse.java
@@ -15,7 +15,7 @@ public class OrderComfirmationResponse {
 	private String title;
 	private LocalDateTime date;
 	private String recipient;
-	private Long zipcode;
+	private int zipcode;
 	private String address;
 	private String addressDetail;
 	private String phoneNumber;
